requirements.txt
# QTCR-Net: Quantum Temporal Convolutional Reservoir Network
# Python dependencies

# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Quantum Machine Learning
pennylane>=0.33.0
pennylane-lightning>=0.33.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Machine Learning Utilities
scikit-learn>=1.3.0
tqdm>=4.65.0

# Configuration and Logging
pyyaml>=6.0
tensorboard>=2.13.0

# Jupyter Notebook
jupyter>=1.0.0
ipykernel>=6.23.0
ipywidgets>=8.0.0

# Optional: PennyLane visualization
# pennylane[matplotlib]

# Optional: Additional quantum backends
# pennylane-qiskit>=0.33.0
# pennylane-cirq>=0.33.0
